requirements.txt
click==8.0.4
joblib==1.1.0
lxml==4.8.0
nltk==3.7
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
regex==2022.3.2
six==1.16.0
tqdm==4.63.0
